feat(chat): add chat input bar, session list, and thread components
- Implemented ChatInputBar for sending messages and file attachments.
- Created ChatSessionList to display and manage chat sessions with filtering and deletion functionality.
- Developed ChatThread to render chat messages and guide content with progress indicators.
- Added GuideVersionTimeline for navigating through guide versions.
- Introduced utility functions in chat-utils for handling markdown and date formatting.
