+++ extension/src/clients/webapp-client.js
@@ -25,23 +25,6 @@ export function isWebappAuthError(error) {
   return error.status === 401 || error.status === 403;
 }
 
-export async function ingestAssignment(payload, baseUrl) {
-  const backendBaseUrl = toBaseUrl(baseUrl);
-  const res = await fetch(`${backendBaseUrl}/api/ingest-assignment`, {
-    method: "POST",
-    credentials: "include",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(payload),
-  });
-
-  if (!res.ok) {
-    const errText = await res.text();
-    throw new Error(`Ingest failed (${res.status}): ${errText}`);
-  }
-
-  return res.json();
-}
-
 export async function createChatSession({ payload, userId }, baseUrl) {
   const backendBaseUrl = toBaseUrl(baseUrl);
   const body = { payload };
@@ -106,36 +89,5 @@ export async function getAssignmentGuideStatus(
     throw new Error(
       `assignment-guide-status returned non-JSON: ${rawBody.slice(0, 300)}`,
     );
-  }
-}
-
-export async function runAgent({ assignmentUuid, payload, pdfFiles }, baseUrl) {
-  const backendBaseUrl = toBaseUrl(baseUrl);
-
-  const res = await fetch(`${backendBaseUrl}/api/run-agent`, {
-    method: "POST",
-    credentials: "include",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({
-      assignment_uuid: assignmentUuid,
-      payload: {
-        ...payload,
-        assignment_uuid: assignmentUuid,
-      },
-      pdf_text: "",
-      pdf_files: pdfFiles,
-    }),
-  });
-
-  const rawBody = await res.clone().text();
-
-  if (!res.ok) {
-    throw new Error(`run-agent failed (${res.status}): ${rawBody.slice(0, 300)}`);
-  }
-
-  try {
-    return JSON.parse(rawBody);
-  } catch {
-    throw new Error(`run-agent returned non-JSON: ${rawBody.slice(0, 300)}`);
   }
 }
