+++ extension/src/shared/constants/widget.js
@@ -1,8 +1,4 @@
 /**
  * Shared widget element IDs.
  */
-export const WIDGET = {
-  CONTAINER_ID: "headstart-ai-widget",
-  MODAL_ID: "headstart-ai-modal",
-  OVERLAY_ID: "headstart-ai-overlay",
-};
+export const WIDGET = {};
