+++ webapp/src/app/globals.css
@@ -8,9 +8,10 @@
 @theme inline {
   --color-background: var(--background);
   --color-foreground: var(--foreground);
-  --font-sans: var(--font-body);
+  --font-sans: var(--font-sans-stack);
   --font-body: var(--font-body);
   --font-heading: var(--font-heading);
+  --font-mono: var(--font-mono-stack);
   
   --color-brand-blue: var(--brand-blue);
   --color-brand-crimson: var(--brand-crimson);
@@ -60,6 +61,14 @@
 
 :root {
   --radius: 0.625rem;
+  --font-apple-sans: -apple-system, BlinkMacSystemFont, "SF Pro Text", "SF Pro Display",
+    "Helvetica Neue", Helvetica, Arial, sans-serif;
+  --font-apple-mono: "SF Mono", "SFMono-Regular", ui-monospace, Menlo, Monaco, Consolas,
+    "Liberation Mono", "Courier New", monospace;
+  --font-sans-stack: var(--font-apple-sans), var(--font-body), sans-serif;
+  --font-markdown-stack: var(--font-apple-sans), var(--font-lexend), var(--font-body),
+    sans-serif;
+  --font-mono-stack: var(--font-apple-mono), var(--font-space-mono), monospace;
   
   /* Brand Colors */
   --brand-blue: #0051BA;
@@ -166,6 +175,10 @@
   }
   body {
     @apply bg-background text-foreground;
+  }
+
+  :where(code, kbd, samp, pre) {
+    font-family: var(--font-mono-stack);
   }
 }
